move version and environment into percy folder
