+++ lib/percy/environment.rb
@@ -1,5 +1,3 @@
-require_relative './version'
-
 module Percy
   def self.client_info
     "percy-capybara/#{VERSION}"
