+++ lib/percy/version.rb
